+++ Dash/DHDocsetDownloader.m
@@ -34,6 +34,7 @@ static id singleton = nil;
                            [DHFeed feedWithFeed:@"Akka.xml" icon:@"akka" aliases:@[@"scala"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Android.xml" icon:@"android" aliases:@[@"java"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Angular.xml" icon:@"angular" aliases:@[@"google angularjs", @"google angular.js", @"angular2", @"angular 2", @"angularjs 2", @"angular.js 2", @"google angularts", @"angular.io angular for typescript angular for ts angular.typescript angular.ts angulartypescript", @"google angular.ts", @"angular2", @"angular 2", @"angularts 2", @"angular.ts 2"] doesNotHaveVersions:NO],
+                           [DHFeed feedWithFeed:@"AngularJS.xml" icon:@"angularjs" aliases:@[@"google angularjs", @"google angular.js", @"angular", @"angularjs", @"angular.js"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Ansible.xml" icon:@"ansible" aliases:nil doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Apache_HTTP_Server.xml" icon:@"apache" aliases:@[@"httpd"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Appcelerator_Titanium.xml" icon:@"titanium" aliases:@[@"Appcelerator Platform"] doesNotHaveVersions:NO],
@@ -107,11 +108,13 @@ static id singleton = nil;
                            [DHFeed feedWithFeed:@"Ionic.xml" icon:@"ionic" aliases:@[@"ionic framework"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"iOS.xml" icon:@"iphone" aliases:@[@"iphone", @"ipad", @"xcode", @"apple", @"cocoa touch", @"objective-c", @"objc", @"swift"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Jasmine.xml" icon:@"jasmine" aliases:@[@"jasminejs", @"jasmine.js"] doesNotHaveVersions:NO],
-                           [DHFeed feedWithFeed:@"Java_EE6.xml" icon:@"jee6" aliases:@[@"javaee6"] doesNotHaveVersions:YES],
-                           [DHFeed feedWithFeed:@"Java_EE7.xml" icon:@"jee7" aliases:@[@"javaee7"] doesNotHaveVersions:YES],
+                           [DHFeed feedWithFeed:@"Java_EE6.xml" icon:@"jee6" aliases:@[@"javaee6", @"jee6"] doesNotHaveVersions:YES],
+                           [DHFeed feedWithFeed:@"Java_EE7.xml" icon:@"jee7" aliases:@[@"javaee7", @"jee7"] doesNotHaveVersions:YES],
+                           [DHFeed feedWithFeed:@"Java_EE8.xml" icon:@"jee8" aliases:@[@"javaee8", @"jee8"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"Java_SE6.xml" icon:@"java" aliases:@[@"javase6"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"Java_SE7.xml" icon:@"java" aliases:@[@"javase7"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"Java_SE8.xml" icon:@"java" aliases:@[@"javase8"] doesNotHaveVersions:YES],
+                           [DHFeed feedWithFeed:@"Java_SE9.xml" icon:@"java" aliases:@[@"javase9"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"JavaFX.xml" icon:@"javafx" aliases:nil doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"JavaScript.xml" icon:@"javascript" aliases:@[@"mdn", @"mozilla developer network", @"dom events", @"canvas", @"js"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"Jekyll.xml" icon:@"jekyll" aliases:@[@"jekyllrb jekyll.rb jekyll ruby"] doesNotHaveVersions:NO],
@@ -146,10 +149,7 @@ static id singleton = nil;
                            [DHFeed feedWithFeed:@"NodeJS.xml" icon:@"nodejs" aliases:@[@"node.js"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"NumPy.xml" icon:@"numpy" aliases:@[@"scipy", @"sci.py", @"num.py", @"python"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"OCaml.xml" icon:@"ocaml" aliases:nil doesNotHaveVersions:NO],
-                           [DHFeed feedWithFeed:@"OpenCV_C.xml" icon:@"cvc" aliases:nil doesNotHaveVersions:NO],
-                           [DHFeed feedWithFeed:@"OpenCV_C++.xml" icon:@"cvcpp" aliases:nil doesNotHaveVersions:NO],
-                           [DHFeed feedWithFeed:@"OpenCV_Java.xml" icon:@"cvj" aliases:nil doesNotHaveVersions:NO],
-                           [DHFeed feedWithFeed:@"OpenCV_Python.xml" icon:@"cvp" aliases:nil doesNotHaveVersions:NO],
+                           [DHFeed feedWithFeed:@"OpenCV.xml" icon:@"opencv" aliases:nil doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"OpenGL_2.xml" icon:@"gl2" aliases:@[@"opengl"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"OpenGL_3.xml" icon:@"gl3" aliases:@[@"opengl"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"OpenGL_4.xml" icon:@"gl4" aliases:@[@"opengl glsl openglsl"] doesNotHaveVersions:YES],
@@ -202,7 +202,7 @@ static id singleton = nil;
                            [DHFeed feedWithFeed:@"Stylus.xml" icon:@"stylus" aliases:@[@"nodejs", @"node.js"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Susy.xml" icon:@"susy" aliases:nil doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"SVG.xml" icon:@"svg" aliases:@[@"mdn", @"mozilla developer network"] doesNotHaveVersions:YES],
-                           [DHFeed feedWithFeed:@"Swift.xml" icon:@"swift" aliases:@[@"leopard", @"snow leopard", @"lion", @"mountain lion", @"mavericks", @"yosemite", @"10.10", @"10.8", @"10.6", @"mac osx", @"10.7", @"10.9", @"10.5", @"xcode", @"apple", @"cocoa", @"objective-c", @"objc", @"iphone", @"ipad", @"cocoa touch"] doesNotHaveVersions:NO],
+                           [DHFeed feedWithFeed:@"Swift.xml" icon:@"swift" aliases:@[@"leopard", @"snow leopard", @"lion", @"mountain lion", @"mavericks", @"yosemite", @"10.10", @"10.8", @"10.6", @"mac osx", @"10.7", @"10.9", @"10.5", @"xcode", @"apple", @"cocoa", @"objective-c", @"objc", @"iphone", @"ipad", @"cocoa touch"] doesNotHaveVersions:YES],
                            [DHFeed feedWithFeed:@"Symfony.xml" icon:@"symfony" aliases:@[@"php"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Tcl.xml" icon:@"tcl" aliases:@[@"tcl/tk", @"tcltk", @"tcl tk"] doesNotHaveVersions:NO],
                            [DHFeed feedWithFeed:@"Tornado.xml" icon:@"tornado" aliases:@[@"python", @"tornado web server"] doesNotHaveVersions:NO],
@@ -233,7 +233,7 @@ static id singleton = nil;
     for(NSDictionary *feedDictionary in savedFeeds)
     {
         DHFeed *savedFeed = [DHFeed feedWithDictionaryRepresentation:feedDictionary];
-        if([@[@"http://kapeli.com/feeds/OS_X.xml", @"http://kapeli.com/feeds/watchOS.xml", @"http://kapeli.com/feeds/iOS.xml", @"http://kapeli.com/feeds/tvOS.xml", @"http://kapeli.com/feeds/Jade.xml", @"http://kapeli.com/feeds/Angular.dart.xml", @"http://kapeli.com/feeds/AngularJS.xml", @"http://kapeli.com/feeds/AngularTS.xml", @"http://kapeli.com/feeds/Gradle_Groovy_API.xml", @"http://kapeli.com/feeds/XUL.xml"] containsObject:savedFeed.feedURL])
+        if([@[@"http://kapeli.com/feeds/OS_X.xml", @"http://kapeli.com/feeds/watchOS.xml", @"http://kapeli.com/feeds/iOS.xml", @"http://kapeli.com/feeds/tvOS.xml", @"http://kapeli.com/feeds/Jade.xml", @"http://kapeli.com/feeds/Angular.dart.xml", @"http://kapeli.com/feeds/Swift.xml", @"http://kapeli.com/feeds/AngularTS.xml", @"http://kapeli.com/feeds/Gradle_Groovy_API.xml", @"http://kapeli.com/feeds/XUL.xml", @"http://kapeli.com/feeds/OpenCV_C.xml", @"http://kapeli.com/feeds/OpenCV_C++.xml", @"http://kapeli.com/feeds/OpenCV_Java.xml", @"http://kapeli.com/feeds/OpenCV_Python.xml"] containsObject:savedFeed.feedURL])
         {
             if(savedFeed.installed)
             {
@@ -249,7 +249,7 @@ static id singleton = nil;
             savedFeed.installed = NO;
             savedFeed.installedVersion = nil;
         }
-        if([@[@"http://kapeli.com/feeds/OS_X.xml", @"http://kapeli.com/feeds/Jade.xml", @"http://kapeli.com/feeds/Apple_Guides_and_Sample_Code.xml", @"http://kapeli.com/feeds/Angular.dart.xml", @"http://kapeli.com/feeds/AngularJS.xml", @"http://kapeli.com/feeds/AngularTS.xml", @"http://kapeli.com/feeds/Gradle_Groovy_API.xml", @"http://kapeli.com/feeds/XUL.xml"] containsObject:savedFeed.feedURL])
+        if([@[@"http://kapeli.com/feeds/OS_X.xml", @"http://kapeli.com/feeds/Jade.xml", @"http://kapeli.com/feeds/Apple_Guides_and_Sample_Code.xml", @"http://kapeli.com/feeds/Angular.dart.xml", @"http://kapeli.com/feeds/AngularTS.xml", @"http://kapeli.com/feeds/Gradle_Groovy_API.xml", @"http://kapeli.com/feeds/XUL.xml", @"http://kapeli.com/feeds/OpenCV_C.xml", @"http://kapeli.com/feeds/OpenCV_C++.xml", @"http://kapeli.com/feeds/OpenCV_Java.xml", @"http://kapeli.com/feeds/OpenCV_Python.xml"] containsObject:savedFeed.feedURL])
         {
             continue;
         }
@@ -705,13 +705,18 @@ static id singleton = nil;
     else if([feed.feedURL isEqualToString:@"http://kapeli.com/feeds/Apple_API_Reference.xml"])
     {
         title = @"Apple API Reference";
-        message = @"To install the Apple API Reference docset you need to:\n\n1. Use Dash for macOS to install the Apple API Reference docset from Preferences > Downloads\n2. Go to Preferences > Docsets, right click the Apple API Reference docset and select \"Generate iOS Compatible Docset\"\n3. Transfer the resulting docset using iTunes File Sharing";
+        message = @"To install the Apple API Reference docset you need to:\n\n1. Use Dash for macOS to install the Apple API Reference docset from Preferences > Downloads\n2. Go to Preferences > Docsets, right click the Apple API Reference docset and select \"Generate iOS Compatible Docset\"\n3. Transfer the resulting docset using iTunes File Sharing or AirDrop";
+    }
+    else if([feed.feedURL isEqualToString:@"http://kapeli.com/feeds/Swift.xml"])
+    {
+        title = @"Apple API Reference";
+        message = @"There is no Swift docset. The documentation for Swift can be found inside the Apple API Reference docset. Install it instead.";
     }
     else if([@[@"http://kapeli.com/feeds/OS_X.xml", @"http://kapeli.com/feeds/macOS.xml", @"http://kapeli.com/feeds/watchOS.xml", @"http://kapeli.com/feeds/iOS.xml", @"http://kapeli.com/feeds/tvOS.xml"] containsObject:feed.feedURL])
     {
         title = @"Apple API Reference";
         NSString *name = [[[feed.feedURL lastPathComponent] stringByDeletingPathExtension] stringByReplacingOccurrencesOfString:@"_" withString:@" "];
-        message = [NSString stringWithFormat:@"There is no %@ docset. The documentation for %@ can be found inside the Apple API Reference docset. \n\nTo install the Apple API Reference docset you need to:\n\n1. Use Dash for macOS to install the docset from Preferences > Downloads\n2. Go to Preferences > Docsets, right click the Apple API Reference docset and select \"Generate iOS-compatible Docset\"\n3. Transfer the resulting docset using iTunes File Sharing", name, name];
+        message = [NSString stringWithFormat:@"There is no %@ docset. The documentation for %@ can be found inside the Apple API Reference docset. \n\nTo install the Apple API Reference docset you need to:\n\n1. Use Dash for macOS to install the docset from Preferences > Downloads\n2. Go to Preferences > Docsets, right click the Apple API Reference docset and select \"Generate iOS-compatible Docset\"\n3. Transfer the resulting docset using iTunes File Sharing or AirDrop", name, name];
     }
     
     if(title && message)