+++ Dash/DHDocsetTransferrer.m
@@ -209,7 +209,7 @@ static id singleton = nil;
 
 - (NSAttributedString *)descriptionForEmptyDataSet:(UIScrollView *)scrollView {
     
-    NSString *text = @"You can transfer docsets using iTunes File Sharing.\n\nFor best results, docsets that are available for download should always be downloaded instead of transferred.";
+    NSString *text = @"You can transfer docsets using iTunes File Sharing or AirDrop.\n\nFor best results, docsets that are available for download should always be downloaded instead of transferred.";
     NSMutableParagraphStyle *paragraph = [NSMutableParagraphStyle new];
     paragraph.lineBreakMode = NSLineBreakByWordWrapping;
     paragraph.alignment = NSTextAlignmentCenter;
@@ -232,7 +232,7 @@ static id singleton = nil;
 
 - (BOOL)emptyDataSetShouldAllowScroll:(UIScrollView *)scrollView
 {
-    return YES;
+    return NO;
 }
 
 - (CGFloat)spaceHeightForEmptyDataSet:(UIScrollView *)scrollView
@@ -240,9 +240,9 @@ static id singleton = nil;
     return 24;
 }
 
-- (CGPoint)offsetForEmptyDataSet:(UIScrollView *)scrollView
+- (void)emptyDataSetWillAppear:(UIScrollView *)scrollView
 {
-    return CGPointMake(0, 16);
+    [self.tableView setContentOffset:CGPointMake(0, -self.tableView.contentInset.top)];
 }
 
 - (void)emptyDataSetDidTapButton:(UIScrollView *)scrollView