+++ Dash/DHPreferences.h
@@ -20,6 +20,9 @@
 @interface DHPreferences : UITableViewController
 
 @property (assign) IBOutlet UISwitch *updatesSwitch;
+@property (assign) IBOutlet UISwitch *alphabetizingSwitch;
+@property (assign) IBOutlet UITableViewCell *updatesCell;
+@property (assign) IBOutlet UITableViewCell *alphabetizingCell;
 @property (assign) BOOL didSetUpdateLabelBefore;
 
 - (IBAction)updatesSwitchValueChanged:(id)sender;