+++ Pods/GZIP/GZIP/GZIP.h
@@ -6,4 +6,4 @@ FOUNDATION_EXPORT double GZIPVersionNumber;
 //! Project version string for GZIP.
 FOUNDATION_EXPORT const unsigned char GZIPVersionString[];
 
-#import <GZIP/NSData+GZip.h>
+#import <GZIP/NSData+GZIP.h>