+++ Pods/HockeySDK/HockeySDK-iOS/HockeySDK.embeddedframework/HockeySDK.framework/Headers/BITCrashDetails.h
@@ -36,27 +36,27 @@
 /**
  *  UUID for the crash report
  */
-@property (nonatomic, readonly, strong) NSString *incidentIdentifier;
+@property (nonatomic, readonly, copy) NSString *incidentIdentifier;
 
 /**
  *  UUID for the app installation on the device
  */
-@property (nonatomic, readonly, strong) NSString *reporterKey;
+@property (nonatomic, readonly, copy) NSString *reporterKey;
 
 /**
  *  Signal that caused the crash
  */
-@property (nonatomic, readonly, strong) NSString *signal;
+@property (nonatomic, readonly, copy) NSString *signal;
 
 /**
  *  Exception name that triggered the crash, nil if the crash was not caused by an exception
  */
-@property (nonatomic, readonly, strong) NSString *exceptionName;
+@property (nonatomic, readonly, copy) NSString *exceptionName;
 
 /**
  *  Exception reason, nil if the crash was not caused by an exception
  */
-@property (nonatomic, readonly, strong) NSString *exceptionReason;
+@property (nonatomic, readonly, copy) NSString *exceptionReason;
 
 /**
  *  Date and time the app started, nil if unknown
@@ -71,14 +71,14 @@
 /**
  *  Operation System version string the app was running on when it crashed.
  */
-@property (nonatomic, readonly, strong) NSString *osVersion;
+@property (nonatomic, readonly, copy) NSString *osVersion;
 
 /**
  *  Operation System build string the app was running on when it crashed
  *
  *  This may be unavailable.
  */
-@property (nonatomic, readonly, strong) NSString *osBuild;
+@property (nonatomic, readonly, copy) NSString *osBuild;
 
 /**
  *  CFBundleShortVersionString value of the app that crashed
@@ -86,12 +86,12 @@
  *  Can be `nil` if the crash was captured with an older version of the SDK
  *  or if the app doesn't set the value.
  */
-@property (nonatomic, readonly, strong) NSString *appVersion;
+@property (nonatomic, readonly, copy) NSString *appVersion;
 
 /**
  *  CFBundleVersion value of the app that crashed
  */
-@property (nonatomic, readonly, strong) NSString *appBuild;
+@property (nonatomic, readonly, copy) NSString *appBuild;
 
 /**
  *  Identifier of the app process that crashed