+++ Pods/HockeySDK/HockeySDK-iOS/HockeySDK.embeddedframework/HockeySDK.framework/Headers/BITCrashManager.h
@@ -37,9 +37,9 @@
 
 
 /**
- * Custom block that handles the alert that prompts the user whether he wants to send crash reports
+ * Custom block that handles the alert that prompts the user whether they want to send crash reports
  */
-typedef void(^BITCustomAlertViewHandler)();
+typedef void(^BITCustomAlertViewHandler)(void);
 
 
 /**
@@ -340,7 +340,7 @@ typedef NS_ENUM(NSUInteger, BITCrashManagerUserInput) {
 
 /**
  Lets you set a custom block which handles showing a custom UI and asking the user
- whether he wants to send the crash report.
+ whether they want to send the crash report.
  
  This replaces the default alert the SDK would show!
  
@@ -350,7 +350,7 @@ typedef NS_ENUM(NSUInteger, BITCrashManagerUserInput) {
  In addition to this you should always ask your users if they agree to send crash reports, send them
  always or not and return the result when calling `handleUserInput:withUserProvidedCrashDescription`.
  
- @param alertViewHandler A block that is responsible for loading, presenting and and dismissing your custom user interface which prompts the user if he wants to send crash reports. The block is also responsible for triggering further processing of the crash reports.
+ @param alertViewHandler A block that is responsible for loading, presenting and and dismissing your custom user interface which prompts the user if they want to send crash reports. The block is also responsible for triggering further processing of the crash reports.
  
  @warning This is not available when compiled for Watch OS!